TIP-221: Move ProductTemplateMediaManager
